Split reactive stories
